Updating speed assignment code to convert value to mph

+++ onebusaway-nyc-transit-data/src/main/java/org/onebusaway/nyc/transit_data/model/NycQueuedInferredLocationBean.java
@@ -69,7 +69,7 @@ public class NycQueuedInferredLocationBean implements Serializable {
 
 	private double bearing;
 
-	private short speed;
+	private BigDecimal speed;
 	
 	// Fields from TDS
 	
@@ -319,11 +319,11 @@ public class NycQueuedInferredLocationBean implements Serializable {
 		this.lastLocationUpdateTime = lastLocationUpdateTime;
 	}
 
-	public short getSpeed() {
+	public BigDecimal getSpeed() {
 		return speed;
 	}
 
-	public void setSpeed(short speed) {
+	public void setSpeed(BigDecimal speed) {
 		this.speed = speed;
 	}
